Stop routing session events via the dap store
This cleans up a bunch of indirection and will make it easier to
show the session building state in the debugger terminal
